MB-66250: Add storage min quota threshold
When new index is created, if it the first of its type,
the quota allotted to it will be very small. The minimum
and mandatory quota computation will lag by 10s. To
avoid memory starvation, inject a minimum quota.
This minimum will let the index build progress and
as the index gets larger, the minimum and mandatory
will become sufficient to sustain the index. This
injected minimum is phased out linearly over 1 minute.

Change-Id: I8399a144ce190170c18fe735a6c4c2728ce84290