MB-62907 Handle race condition between Dequeue and Destroy
When the last index of a stream, keyspace is dropped while
flush is in progress, indexer will soft delete the index,
updates the flusher maps and destroys the queue. The soft
deleting logic has been added as a part of MB-57188.

In mutation queue Destroy(), the mutation queue will drain
all mutations and set the head.next.mutation to nil. However,
it can race with Dequeue() method due to the change in MB-57188.

Because of this, if Destroy() happens and dequeueUptoSeqno wakes
up and reads head.next.mutation, then the mutation can be nil
and the subsequent access to m.meta.seqno will panic.

To avoid the panic, mutation queue will do a nil check first
and only then access the meta.seqno. Also, the dequeue method
would exit normally as if the flush is done. Indexer and other
components like timekeeper will gracefully handle the exit.

Change-Id: Ibaa744a09024f59feead0f5bd920f65a8bb4ec25