MB-66799 Handle race between DequeueUptoSeqno and DequeueSingleElement
During Destroy(), there can exist a race between DequeueUptoSeqno
and DequeueSingleElement where DequeueUptoSeqno will set the
head.next.mutation = nil and DequeueSingleElement will read the
same head.next.mutation.

This will lead to m.Size() triggering a panic in DequeueSingleElement
as head.next.mutation is nil. This patch avoids the fix by adding
a nil check to DequeueSingleElement. As the return value of this
method can be "nil" in 2 cases - When the queue is empty and also
when there is a race, an additional return value 'queueEmpty' is
added.

Change-Id: I7e12a6f7e70169e1247efe85b8aa532c374d867e