MB-66478 Add all indexes being repaired to clonedCandidates
The current logic for adding an index to clonedCandidates has a bug
because of which only the first index that is being added to proxy
is considered under optional indexes and the rest all are ignored
under optional indexes. If planner could not a solution satisfying
resource constraints, it will remove all optional indexes from the
cluster.

However due to the current bug, many indexes will still remain in
the solution inspite of planner trying to remove optional indexes
as those indexes were not added to optional indexes. To avoid this
case, this patch fixes the issue by adding all indexes being repaired
as optional indexes.

Initial version of the code had a check for eligibility of index
before adding to optional indexes (i.e. clonedCandidates list)
but that check is unnecessary as a shard will be considered as
eligible for movement only if all individual indexes of the shard
are eligible. Therefore, even if an index getting repaired is
placed on a shard that is not eligible for movement, planner will
not consider the shard eligible for movement.

Change-Id: Ie87b68fdf9bfc3039ce65571cd15637feb678b24