MB-66656: Create min num of shards before reusing shard in that category
There can be a skew where the user creates "category 1" indexes till the
MinShardCapacity per node is reached. Meaning each Index of "category 1"
goes to new shard. And if the subsequent index creations are of an index
belonging to "category 2", only 1 shard can be created for "category 2"
and continously reused. This can cause all the indexes of category 2 to
be saturated on one node and always moved together, since they share the
same shard.

To prevent this and allow for a better distribution of indexes on
different shards, a minimum capacity of each category of shard is
created before reusing. This minimum capacity overrides even if the
tenant level shard capacity and perNode minShardCapacity is hit.

Change-Id: I34b96d1802c4ccfd32e93e68c7280cb24e4cc435