MB-66999: Do not extend a slot to support with Nodes in ShardDealer
bg: whenever with nodes clause is used for an index creation, only the
local metadata of the target nodes is fetched and passed for planning.
Whereas ShardDealer assumes every information passed to it is the global
metadata i.e. metadata from all nodes in the cluster. This global
metadata is used to decide whether to extend or reuse or create a new
alternate shardId.

problem: since this is partial information, Shard dealer can make a
decision of extending a slot on one of the target nodes even though the
new extended slot-replicaId is present in the cluster (just not on the
target nodes) and Hence one alternate shardId(slotID-replicaID) is now
present on more than 1 node, breaking the invariant that one alternate
shardID only maps to 1 shard pair.

fix: For target nodes whenever a slotID cant be reused, do not check if
any slot can be extended since the information passed to Shard Dealer is
not global. and hence force creation of a new alternate shard id.

Change-Id: Ic9f0225311ae6ad298ad4e970d27b40022697b5b