MB-67033: Compute resident ratio for bhive
Resident Ratio is computed using numRecMem and numRecDisk.
numRecMem is the records in memory for graph and flush buffer.
numRecDisk is the records on disk for graph and flush buffer.
Magma will not contribute to resident raito.  The Magma
memory used is part of MemSzIndex (fixed portion of index
that do not scale proportional with mutation and scan traffic).

Change-Id: Ifefc6704c815439a81023539c8212324e1df2611