MB-67157 Fix resident ratio calculation for shard proxy
For a plasma index, the resident ratio is calculated by the
formula: recsInMem/(recsInMem + recsOnDisk).

In planner, recsInMem is being read as ActualRecsInMem and
recsOnDisk is being read as TotalRecords. In the current
implementation, the resident ratio of shard proxy is calculated
by the formula: ActualRecsInMem / TotalRecords.

This is incorrect because if the index is 90% resident, then the
resident ratio of shard proxy will be wrongly estimated as 900%.
This patch fixes the calculation of resident ratio for shard proxy.

Change-Id: Iafd2bc889e57cd528e4f156d99bebff417a3cf51