MB-65942: Dont fail rebalance during Validate Train list
Due to an invalid config of train list size, the create index during
rebalance can fail. This can be caused by cases such as bucket flush
before rebalance.
Rather than failing rebalance due to the vector train list check, the TT
is promoted to TransferTokenCommit and the Index remains in error state
on the dest node.

Change-Id: I806454a09fbdcb6fee127fa3252f42eedd50746a