MB-65842: Add config for plasma hole finder
With holeFinderMinEmptyPercent, the aim is to avoid
hole unnecessary finding attempts. If 10% of the
pages are empty, there is a good chance that there
are holes. If the empty pags are sparsely distributed
then, we may still not be able to find any suitable
holes

With holeFinderMinNumPages, the aim is to avoid
hole cleaning when the index is very small. With
a minimum of 10 pages, with many indexes, if all
pages are empty, the overhead adds up to a
relatively small number.

Change-Id: Iaa2a76147353feff0db022be60d1d7fa023cc226