MB-65161: Update plasma global hole punch granularity in slice
Plasma uses a global configurable for the reclaim block
size. This used to be set once during warmup, with no
way to change it without restarting the indexer. Now,
the config is set in updatePlasmaConfig. Note that the
global value is set once per slice and all plasma
indexes' LSS will pickup this new value for hole punch.

Change-Id: I996a2cc95d7b46f29679dcdf7849e51773408237