MB-66878 Handle drop when training in progress gracefully
When index training is in progress and drop is received, current
code would wait for training to finish and will return the
response back to caller.

During rebalance CancelTask(), there is an issue with this sync
response. ns_server has a strict timeout of 60sec for RPC calls.
If dropIndex takes more than 60seconds, then the RPC connection
between index service and ns_server will be lost. Because of this,
further requests to rebalance will fail as there is no RPC connection.

Another issue is that if dropIndex takes more than 30min, then ns_server
rotates the credentials. For new credentials, index service has to
restart. This causes service outage and rebalance failures.

This patch addresses the issue by not blocking the dropIndex call
from training to finish - only during rebalance. Instead, the error
will be returned back to the caller. For a user initiated drop,
the behaviour is not changed.

Change-Id: Ib8996098e850c7dffd3f64d4e49fc202b70db5b2