MB-100 remove parallel calls to test to reduce memory load
shard dealer test seems to be failing in CI due to OOM kill
and somehow the memory usage is rising to 12GB. full run of
the test on local setup is not leading to such high memory
usage so not sure what can cause this.
suspect is that due to parallel calls, individual test does
not end until all tests finish and hence memory is not
getting freed.

Change-Id: I1b52e59a5d148b7ef623dbc3454e761de23cf495
Signed-off-by: Dhruvil Shah <dsdruvil8@gmail.com>