MB-67260 Respond with success instances only after training is done for all instances in the batch
In a batch of indexes that are being trained, consider the following
case:

a. Index idx_1 has enough vectors from sample and finished training
b. Index idx_2 does not have enough vectors and needs re-sampling
c. In the current code, re-sampling for idx_2 will be attempted but
   success message for idx_1 will be sent back to indexer
d. Indexer will start build for idx_1
e. If idx_2 finishes training and attempts build, build will fail
   with "Build Already In progress" error for this keyspace.
f. To prevent such a case, this patch updates the code so that
   success message must be sent only after all retries are finished
   for the batch.
g. Error messages are sent as and when training fails to make sure
   indexer book-keeping is updated so that other blocking operations
   like dropIndex will finish without waiting for entire training
   batch finishes

Change-Id: Iec9f0e660942f44343919fb081bc9fe61fdfd8be