MB-66361 Increase the size of closeMutQueue to 2
As there can be 2 writers to this channel (doReceive and
genServer) while there is only one listener (DequeueMutations),
increasing the channel size to "2" ensures that both writers
can write to the channel even if the listener exits.

TL;DR
In a rare sequence of events, it is possible that the DCP
connection can be terminated while DequeueMutations() is sending
buffer ack and doReceive is trying to receive mutations and genServer
is yet to process dfCmdClose message. In such a case,

a. DequeuMutations() exits and closes the dequeueDoneCh
b. doReceive exits and writes to closeMutQueue
c. genServer when processing dfCmdClose will try to write again to
   closeMutQueue but since the channel size is only "1" and the
   only listener on this channel exited, genServer will be blocked
   permanently leading to a deadlock.

To avoid such a case, the size of closeMutQueue is increased to 2
so that even if listener exits first, both writers can successfully
write to the channel and remain unblocked. As listener has already
existed, there is no effect of writes in closeMutQueue channel.

Change-Id: Iffae8ce98cd120569130d9524aa7c91eb846082f