MB-64279: Only invalidate bucket or collection if deleted/dropped
bg: earlier the ClusterInfoCacheLite's validate Bucket and CollectionID
returns invalid even when it encountered any errors related to CICL or
if the request timed out (can be due to lost CBAUTH connection). This
can cause deletion of all metadata for indexes lying on that keyspace.

fix: Return invalid only for confirmed deleted/dropped keyspace cases.
For any other error - that is for cases when the check itself could not
be made - consider it a valid bucket/collection. For false positives,
the check is passed down, so when making the call for stream etc error
of missing bucket or keyspace can be handled, rather than deleting the
metadata of indexes.

Change-Id: I729b0b2e89363c71462247710029601a8c0d60cf