MB-67808 Add checkForLostPartitions in monitorItemsCount
This check compares the number of partitions for each index
instance and logs a fatal message if the number of partitions
in the cluster is not same as the number of partitions in
the index definition.

The check for lost partitions succeeds only if there is no
error in retrieving the index metadata from the nodes. Otherwise,
the information can be partial and it can lead to false
positives.

Change-Id: Iea109bc6a26dbf7b9ff5fa8e5ced3d7b289d6aa3