MB-67808 Use defnId instead of IndexName for matching replicas
As the REST call happens to various nodes at different points in time,
it is possible for one node to return the timestamped counts info
for one defnId while the other nodes return timestamped counts for
same index name but with different defnId. Ideally, both should not
be compared. Hence, using defnId to sort the replicas instead of
index name

Change-Id: Ied5041035fed7ca121eca1ddb1ef5d38c6de2599