MB-67331 Initialise endpoints if the KeyPartn container has zero partitions
During rebalance, it is possible that all partitions of real
instance are pruned leaving zero partitions in the real instance.
At this time, if there is a merge of partitions, then the real
instance would not contian any endpoints. This will lead to
projector not streaming mutations for real instance causing
index corruption.

Change-Id: I2fd9ea6eef16ac47dba6c877b39ca84d97d767d9