MB-66655: Enable shard dealer during restore
Backup of indexes can contain vector indexes, if the target cluster
version is 8.0 enable shard dealer when the shard affinity can be
maintained.
This is done as in  8.0 shardAffinity and ShardDealer work as a single
config. Shard dealer config won't be exposed/recommended to be changed
by the customer. So If shardAffinity is disabled, shardDealer should be
disabled.

Change-Id: I6363a9a767d6f19552333aa12d255e27ae9468df