[BP] MB-68051: Use sorting function instead of selection sort for buildList
If the pending map is huge as well as the key (keyspace string) is big,
it cause the overall map lookup to take a lot of CPU cycles.
The map look ups are fixed at n^2. Using the sorting function we can
bring down the map lookup and hence CPU usage significantly.

Change-Id: I2bf4475080d56bebb40e8e1329043ef1875286c5