MB-67851 Release excess readers in semaphore
The number of readers acquired and released per slice is
dependent on the readerPerPartition value. Since readersPerPartition
is being reduced here due to "perPartnScanParallelism", we also need
to release the equivalent amount of readers in semaphore. Otherwise,
slice.releaseReaders() will end up releasing less readers than it
acquired (as readersPerPartition is a shared variable) and we will
end up in eventual deadlock.

Change-Id: I4433e840159b166e357d9332300442a65cbac2e8