MB-67945 Update shardIds before recovering codebook
In bhive_slice, update shardIds before recovering codeboook.
For deferred instances, codebook recovery can lead to error
as the codebook won't be available. Current implementation
returns error without updating shardIds. As the error is
expected (due to index being deferred), it will be ignored.
But the shardIds in slice will not be updated.

During shard rebalance, if the shard has to be destroyed, then
indexer waits until all slices on the shard are closed before
destroying the shard. This requires the slice to have valid
shardIds in the metadata. In this case, as that information is
not available, shard got destroyed while slices are active
leading to fatal errors from bhive.

Change-Id: Ib98d2107b357423f4a9fbe02007afc546493d496