MB-68081 Update RState at storage manager after partition merge
The "checkForLostPartitions" code will consider only instances
that are not in RState = RebalPending at storage manager. When
a partitioned index moves to a new node (and only one proxy
moved) to the destination node, then the RState will be updated
at indexer but not at storage manager.

This causes storage manager to skip some partitions of the index
for the "checkForLostPartitions" even if the partitions are active.
This patch fixes the issue by updating storage manager with the
latest index instance map after merge happens.

Change-Id: I89cbd1612ef1fd3ad54d138932446decfab1369e