MB-68105: Calculate minMemory using dataSize for very low resident indexes
For indexes which are very low resident ( <1.0% RR), instead of
estimating minimum memory by scaling the page memory use data size.

fix: for any indexes at less than 1.0% RR estimate the Resident ratio to
be at 0.0% RR. This ensures the datasize is used for min memory
calculation furthermore, the calculation becomes inline with the
previous versions where the resident ratio was truncated to 0.

Change-Id: Id6869f4fdc51f261681c09ba22e80b868288e297