MB-67485: Add maximum retries for replanning during rebalance
Currently there are no maximum limit till the replanning can run for
rebalance. This can mean if there is an unexpected errors that can't
be resolved through more layout planning, the planner can run for
indefinitely long.

fix: add a max upper limit for planner retries only for rebalance. This
limit is configurable.

Change-Id: Ifc24dbb3b0c2324cab9e4e1f62414e45ab54945b