MB-67892 Handle pending build indexes on destination during DCP rebalance
During DCP rebalance, if the destination node has realInst in
pending build state, then current logic only works if the
version of the realInst is 0. Otherwise, DCP rebalance would hang.

E.g., in the following case, without this patch DCP rebalance would
work:

a. Partitioned index created on node n1 and n2 in deferred state
b. Index build happened and indexes on node n1 yet to be built
b. Node n2 is removed and rebalance triggered -> this rebalance
   completes without any issue.

Now, consider the following case:

a. Partitioned index created on node n1 and n2 in deferred state
b. Node n1 is removed and the deferred indexes on n1 moves to n3.
   During this movement, the version of the deferred instances
   is incremented
c. Now, build is issued. Indexes on node n3 yet to be built but
   indexes on node n2 finished building
d. Node n2 is removed and rebalance is triggered. This rebalance
   will remain stuck forever on merge partition.

  DCP rebalance works in scenario-1 but not in scenario-2 because
  in scenario-1, the inst.Version == 0. Because of this indexer
  will set the RState of realInst to REBAL_ACTIVE and merge happens
  without any problem.

  In scenario-2, since inst.Version is incremented due to first
  rebalance, the RState of real instance will be set to
  REBAL_PENDING causing merge to block.

This patch addresses the issue by explicitly tracking realInsts
at the start of rebalance. If any build request comes for
realInst and if it is in the map, then the realInst RState
will be updated to REBAL_ACTIVE instead of REBAL_PENDING

Change-Id: I72fcc4739a9e88312f9d9910e6de366a97d97a4c