MB-67892 Allow pendingBuild indexes as a part of shard rebalance
If there are any pendingBuild instances on the shard (or) vector
instances with training errors, then shard rebalancer will build
them via "buildRecoveredIndexes" API. This means that the indexes
on rebalance source node can be in CREATED/READY state and moved
via shard rebalance. To admit those requests, the checks at
lifecycle manager and indexer are updated to allow build on indexes
in CREATED/READY state.

This is done for 2 reasons:

a. If shard rebalance is moving indexes and build request is received
   for them, then it expects the instances to be built. Therefore,
   instead of rejecting the request, this patch allows to build the
   pending build indexes as a part of shard rebalance
b. In an other case, it is possible that rebalance source node has
   all active instances and rebalance destination node has pendingBuild
   instances. For rebalance to complete successfully, the instances on
   both source and destination has to arrive at same state. This means,
   that the pending build indexes on destination node should be built
   along with the incoming instances. This patch also address the issue
   by including the instances on destination node that are yet to be
   built.

Change-Id: Ie73399e2b2c59e2817fcc5555fa5c401f5adbb4b