MB-67893 Ignore instances already in training during shard rebalance
For shard rebalance initiated builds, if training is already in
progress, then skip them from further processing. Indexer will
build the index again after training is complete.

Otherwise, if vector and non-vector indexes are built in same batch
and in the next request, only non-vector indexes are built by shard
rebalancer, then the non-vector indexes will be built. Later, after
the training is complete, if shard rebalancer attempts to rebuild
the indexes, it will result in an error as shard rebalancer expects
instance to be in READY/CREATED/RECOVERED state but the index will
already be in ACTIVE state.

Change-Id: I5bedf9ea868faa17630019b2f03cb0d39343223d