MB-67893 Do not update partitions after training is done
As the list of partitions are not changing, do not update the
partitions in the meta after training is done. However, perform
a syncUpdate so that the training phase and error are in sync
between indexer and lifecycle manager.

Otherwise, as the partitions are updated async, if the update
happens after some partitions are pruned, then the update will
add back the partitions to lifecycle managers metadata and this
can lead to rebalance failures.

Change-Id: I301ecff42c302628cdc537f3debd6ba5103d2493