MB-68276 Reset state to INDEX_STATE_READY during shard rebalance
If training fails during shard rebalance, then reset the state of
the index to INDEX_STATE_READY (from INDEX_STATE_RECOVERED). This
ensures that rebalance can finish gracefully and index can be built
later. Training can fail during shard rebalance if there are not
enough documents and the instances are in error on both source and
destination nodes.

If the state is not changed, then there will be 2 issues:
  a. Shard rebalancer continues to wait for index to move to READY
     state and that will never happen
  b. The index can not be built using build index if the state is
     not changed

This patch addresses the issue by changing the instance state to
READY after training is done (i.e. for failed instances). For
successful instances, index build is triggered which will change
the state to INITIAL (from RECOVERED).

For non-training errors, rebalance would fail.

Change-Id: I71c257c2ccdf980f857bbf8d0dc0943e42102056