MB-68308 Update error in topology only when asked for
Current code resets error in topology even when the error
update is not asked for. Consider the following case:

a. During rebalance, training error is seen and topology
   is updated
b. After rebalance finishes, RState of the instance is changed
c. During RState update, only updateRState is requested by indexer
d. However, the current code will reset the error to empty string
e. This will cause the index to exist without any error in metadata
f. During next rebalanace, indexer assumes there is no error and
   initiates codebook transfer. Codebook transfer fails with
   codebook does not exist error.

The fix for this issue is to update error in metadata only when
explicitly asked for.

Change-Id: Idaa3b032afb154ef9c28148a283fe2ef11fd5397