MB-68350 Do not skip RState change for pendingBuild indexes
In MB-67892, we added support for building indexes that have
the pendingBuild flag set to true via shard rebalance. However,
we missed updating the RState for such indexes at the end of
rebalance. This causes the newly moved indexes to be in RebalPending
state.

This patch addresses the fix by changing the RState of pending build
indexes at the end of rebalance.

Change-Id: I8b306613290a41f1e7043aa00c40881233b541bc