MB-68308 Prevent go-routine leak in indexer
Indexer can not perform sync update on its own. Sync update
is valid only if indexer waits on respCh. If indexer is not
waiting on respCh, then it is not sync update.

In our case, sync update is set to true in 2 places:
1. updateTopologyOnShardIdChange
2. handleIndexTrainingDone

In upateTopologyOnShardIdChange, a blocking respCh is choosen.
If there is a topology change, then the go-routine in cluster
manager agent will be blocked as there is no listener. This
patch fixes the issue by making it a non-blocking channel.

In handleIndexTrainingDone, a nil respCh is passed. This means,
the go-routine in cluster manager agent remains stuck forever.
This patch removes the syncUpdate = true and makes it false
as indexer is anyways not waiting for the responce channel.

Change-Id: Icb6b1ba8ab72050282400b3e746df49f9dc93ed1