MB-68454 Skip clean up for instances in training phase due to bucket drop
If bucket drop happens while training is in progress, indexer
currently goes ahead and closes the slice. This can lead to a
panic on indexer as the codebook can be reset while training
is happening.

There is already infra to check if the instance is in training
or not. Instead of pruning the instances that are under going
training, all instances on the keyspace are dropped in the
current code.

This patch fixes the issue by skipping the drop for instances
that are in training phase. After training is complete, the
index will be dropped.

Change-Id: I91445f820ef0059ac4aa976110ee51c99eefb689