MB-68535 Update RState of pendingBuild indexes during cleanup
If shard rebalance fails after pendingReady is set to true, then
the RState of the indexes will be updated as a part of cleanup.

For deferred vector instances in error state, this RState update
is skipped. This can lead to a case where the instance is ignored
from further processing during the next set of rebalances and
lead to issues like panic, codebook does not exist, shard already
exists errors.

Change-Id: Id8661e88012519d421e445708839f35ba7d11d0f