MB-68531 Consider training phase instead of training error for codebook transfers
When DDL service manager creates vector indexes, there are
a few cases where a non-deferred vector index can exist in
CREATED phase without any training error. So far shard rebalancer
relied on training error for the transfer of codebook.

If there is no training error, then shard rebalancer expects
to transfer codebook but rebalance would fail as there is no
codebook. It is easy to enter such cases for indexes created
via DDL service manager. One such case is for partitioned indexes
where alter index is performed after first vector index
ended up in error. The altered index will endup in created
state but it will not be built.

Although fixing the DDL service manager can resolve the issue,
it makes more sense to handle this new state of vector instances
(i.e. non-deferred instances in CREATED but not built state)
with shard rebalance so that the fix is generic and does not
rely on error.

However, to rely on TrainingPhase, it has to be populated in the
index usage at planner and then used during codebook path generation
for shard rebalancer.

The following cases are tested with this change for shard rebalance

a. Non-deferred index in training error, index moved to different node.
   Rebalance completed successfully. RState of moved index is RebalActive
b. Deferred index in training error, index moved to different node.
   Rebalance completed successfully, RState of moved index is RebalActive
c. Partitioned index versions for (a) and (b). Rebalance completed
   successfully.
d. Alter index on partitioned and non-partitioned versions when vector
   index is in training error.
   Found https://review.couchbase.org/c/indexing/+/233940 during the
   testing
e. Partitioned index - error on one node and no-error on other node.
   Move the errored out index to non-error node and vice versa
f. Non-errored versions of (a) and (b)

Change-Id: I9c853d911668fa51b20d162930ecb442b47e9393