[BP to 8.0.1] MB-69145 Fix alternate shardId population for alter index
BP from MB-68572

When alter index is performed on an index with alternate shardIds,
the replica alternate shardIds are formed based on the alternate
shardIds in the defn passed to replica repair.

E.g., if replica repair is formed based on the definition containing
partitions 1,2 - then for the definition containing partitions 3,4
alternate shardIds for partns: 1, 2 will be used. This will lead to
the partitions being built with empty alternate shardIds.

Change-Id: I2aa446632fbcf88cc94c0615c3c85e399a8e42f4