[BP 7.6.9] MB-69231: Dont transition TT to Ready when cancel is called during Merge
BP MB-68558

During startShardRecovery, when the indexes are Recovered, and the
function is waiting for Index state as ACTIVE. During the default loop,
all the instances are iterated through, if the instances reach ACTIVE
state but at the time of finishSuccess() the instId is already deleted
from the processedInsts list, there is no guarantee that the merge of
destTokenToMergeOrReady went through. This can lead to the TT to be
transitioned to the next state even though merge won't happen.
If it transitions, the cleanup will not be done by the Destination node

To prevent the Transfer token from moving to the next state, errgroup is
used to capture any errors returned by the merge observer go-routines.
Presence of error in any one of the go-routine signifies that either
rebalance was cancelled or done.
The errorgroup waits for all the go-routine to be done as well checks
for any returned error. For any returned errors, don't proceed further

Change-Id: Ie13a85a4096a025a1e34c5770a24f46801a494d9