[BP] MB-68068 Initialise endpoints if the KeyPartn container has zero partitions
BP MB-67331 to 7.6.8

During rebalance, it is possible that all partitions of real
instance are pruned leaving zero partitions in the real instance.
At this time, if there is a merge of partitions, then the real
instance would not contian any endpoints. This will lead to
projector not streaming mutations for real instance causing
index corruption.

Change-Id: I545b23b938a01ab0a030ac6e3373c25640c0bd29