[BP 8.0.1] MB-69234: Add mechanism to retry security context initialisation on error
BP MB-68943

Currently if any error is encountered while executing the function
inside the sync.Once primitive, the primitive is marked as Done even
though the function inside fails. So when the secuirty context for
client is being initialised if any error occurs, the security context is
Not Set and It is never tried again due to sync.Once being marked as
done.

fix: introduced a new type which is very similar to sync.Once but with a
key difference that it is aware of the result (error) from the function
inside of it, to mark itself done only when it succeeds. For any error
it is not marked done and can be invoked again.

Change-Id: I12bf1395187515edb65917bd4e7fedfb1797eeac