MB-68576 return DDL in progress for bhive graph build
Since Bhive graph build is a CPU intense task, it can lead to CPU saturation when done in conjunction with rebalance. For this, rebalancer already makes rebalance wait till Bhice graph build is completed. Instead of rebalance being stuck for long time (several hours for large indexes), it is better to reject rebalance by considering it as DDL in progress.

Since there isnt a instance-wide GraphBuildPhase like there is a TrainingPhase, I did think of introducing that as it wouldve made this check cleaner and could be used elsewhere. But didnt want to add a new field and all transitions for just this one thing.

Change-Id: If115aefa84f2040f417c1889e792edaaf71aa595