MB-58796: Populate the prometheus per partition stats for lost replica
Similar to diverging replica, the num_lost_replicas and
num_indexes_with_lost_replicas are populated in prometheusMetricsHigh
endpoint.

The stats are only populated if there is atleast one replica loss
observed for the partition. For index partitions having no replica loss
this stat will be absent in the output of prometheus endpoint.

Updated the metrics_metadata.json with the gauge names and info.

Change-Id: I31944209eb9ab890d744648554bd647680834b83