MB-68875: Log a warning and suppress error to storage manager
Whenever shard manager of bhive is not initialised (this can be due to
no Bhive index has been created on node) an error log will be printed.

fix: For the storage not initialised error, print a warning and suprress
the error when returned to the storage manager. Suppressing the error
should not any side effect, since the bhiveShards slice returned would
be nil.

Change-Id: I46d29d66a402aed57086cd76992221959e14ab75