MB-69970 Break Shard affinity for non-replicated indexes on rebalance out
BP MB-64311 to 8.0.3

It is possible that the Proxy replicas are more than the nodes but an
index within a proxy could be accommodated if broken out of the proxy.
e.g.
        Node n1 has s1-r0{i1 (replica 0), i2 (replica 0), i3(replica 0)}
        Node n2 has s1-r1{i1 (replica 1)}
        Node n3 has s1-r2{i1 (replica 2)}

If node n1 is removed, the s1-r0 can't be kept in the cluster, but i2
and i3 will be lost if the entire shard is dropped blindly.

fix: Whenever such a condition is encountered where the entire shard
proxy is getting dropped in dropReplicaIfNecessary, check for any index
definition loss that can occur, if so then place the index definition as
a replica repair and let DCP rebuild happen for that index

Change-Id: I9999aa603a4efff2888b99114950cf3df8e8231c
(cherry picked from commit 0d586758aaf6c92d7cc62cf949519ad9999123a9)