MB-69044 Adds retry for build index if index not found in metadata
Some of the test with defer build true are failing intermittently because Build index started before metadata was updated, thus it got index not found but the create index with defer build true was actually successful.

Fix:
Adding a retry after 1 second to let it get updated in metadata before calling build indexes.

Change-Id: Ia0ddbf64b72753eb0a09ae07a1f4572df068ba8e