MB-68544 Reject merge if source/target instances are in training phase
If source or target index instances are in training phase,
the reject the merge. Indexer will trigger merge after
training is done for those instances (if the keyspace is idle,
then indexer needs to force a merge after training is done
to properly merge the erroneous instances).

After training, if source and target do not agree on the same
training state, then fail rebalance. This can happen in the
following case:

1. Real instance got build request and training initiated
2. A proxy (p1) moved to the node and it started build
3. Real instance did not find any documents and marked
   for training error
4. While proxy p1 training is in progress, another proxy
   p2 started training where both realInst and p2 have
   started to train
5. p1 moves to TRAINING_NOT_STARTED but the training for
   realInst and p2 succeeds now. So, they will be moved
   to TRAINING_COMPLETED state
6. In this case, we can not merge realInst and proxy (p1)
   and rebalance would fail. Although this is a rare race
   condition, it is better to handle it.

A retry of rebalance should fix the issue as the documents
are added between in the middle of training different proxies.

Change-Id: I01c704a9eb0f2431304d426536dab84fc8ca941a